Rename folder from BitmapPainters to Painters
