+++ TagsCloud/TagsCloud.CLI/EntryPoint.cs
@@ -3,8 +3,8 @@ using System.Drawing.Imaging;
 using System.Text;
 using SimpleInjector;
 using TagsCloud.Core;
-using TagsCloud.Core.BitmapPainters;
 using TagsCloud.Core.Layouters;
+using TagsCloud.Core.Painters;
 using TagsCloud.Core.Settings;
 using TagsCloud.Core.TagContainersCreators;
 using TagsCloud.Core.TagContainersCreators.TagsPreprocessors;
