+++ TagsCloud/TagsCloud.Core/Painters/DebugPainter.cs
@@ -2,7 +2,7 @@
 using TagsCloud.Core.Settings;
 using TagsCloud.Core.TagContainersCreators;
 
-namespace TagsCloud.Core.BitmapPainters;
+namespace TagsCloud.Core.Painters;
 
 public class DebugPainter : ITagPainter
 {
