+++ TagsCloud/TagsCloud.Core/Painters/Pallets/MonocolorPallet.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagsCloud.Core.BitmapPainters.Pallets;
+namespace TagsCloud.Core.Painters.Pallets;
 
 public class MonocolorPallet : ITagCLoudPallet
 {
